+++ app/src/scenes/Health/Conseils.js
@@ -1,41 +1,56 @@
-import React from 'react';
 import styled from 'styled-components';
 import H2 from '../../components/H2';
 import TextStyled from '../../components/TextStyled';
 import { listConseils } from './ListConseil';
 import { defaultPaddingFontScale, screenHeight, screenWidth } from '../../styles/theme';
 import AppointmentHeart from '../../components/illustrations/AppointmentHeart';
-import H1 from '../../components/H1';
-
+import ChatBubbles from '../../components/illustrations/Chatbubbles';
+import { View, TouchableOpacity, Text } from 'react-native';
 import { logEvent } from '../../services/logEventsWithMatomo';
 import WrapperContainer from '../../components/WrapperContainer';
 
 const Conseils = ({ navigation }) => {
   return (
     <WrapperContainer title="Santé">
+      <>
+        <H2 color="#4030a5" className="mb-2">
+          Parler avec un professionnel
+        </H2>
+        <H2 color="#000000" className="mb-2">
+          Gratuitement et anonymement
+        </H2>
+        <CategorieContainer
+          onPress={() => {
+            logEvent({
+              category: 'CONTACT',
+              action: 'CONTACT_OPEN',
+              name: 'HEALTH',
+            });
+            navigation.navigate('CONTACT');
+          }}>
+          <IconContainer>
+            <AppointmentHeart size={40} />
+          </IconContainer>
+          <TextContainer>
+            <TextStyled> Prendre un RDV</TextStyled>
+            <TextStyled> avec Doctolib</TextStyled>
+          </TextContainer>
+        </CategorieContainer>
+      </>
+
       <H2 color="#4030a5" className="mb-2">
-        Parler avec un professionnel
+        Témoignages
       </H2>
-      <H2 color="#000000" className="mb-2">
-        Gratuitement et anonymement
-      </H2>
-      <CategorieContainer
-        onPress={() => {
-          logEvent({
-            category: 'CONTACT',
-            action: 'CONTACT_OPEN',
-            name: 'HEALTH',
-          });
-          navigation.navigate('CONTACT');
-        }}>
-        <IconContainer>
-          <AppointmentHeart size={40} />
-        </IconContainer>
-        <TextContainer>
-          <TextStyled> Prendre un RDV</TextStyled>
-          <TextStyled> avec Doctolib</TextStyled>
-        </TextContainer>
-      </CategorieContainer>
+      <View className="border border-[#4030A5] bg-white rounded-md shadow-md my-4 py-4 px-3 mb-8">
+        <TouchableOpacity
+          className="flex-row"
+          onPress={() => {
+            navigation.navigate('TESTIMONIES');
+          }}>
+          <ChatBubbles size={40} className="ml-6" />
+          <Text className="ml-8">Consulter des{'\n'} témoignages</Text>
+        </TouchableOpacity>
+      </View>
       <H2 color="#4030a5">Mes articles conseils</H2>
       <ViewConseilsContainer>
         {listConseils.map((conseil, index) => (