+++ app/src/scenes/Infos/MentionsLegales.js
@@ -7,25 +7,27 @@ import TextStyled from '../../components/TextStyled';
 import { Bold } from '../../components/Articles';
 
 const MentionsLegales = ({ onClose }) => (
-  <WrapperContainer title="Mentions Légales - Application Oz Ensemble" onPressBackButton={onClose}>
+  <WrapperContainer title="Mentions Légales de la plateforme Oz Ensemble" onPressBackButton={onClose}>
     <Spacer size={20} />
-    <H2>Édtieur de l'application</H2>
+    <H2>Édtieur de la Plateforme</H2>
     <Spacer size={30} />
     <P>
-      L'application est éditée par l’Agence régionale de santé (ARS) Île-de-France située{'\u00A0'}: {'\n\n'}
-      Immeuble Le Curve{'\n'}
-      13 Rue du Landy{'\n'}
-      93200 Saint-Denis{'\n'}
-      Téléphone : 01 44 02 00 00
+      L'application est éditée par la Fabrique numérique des ministères sociaux, située{'\u00A0'}: {'\n\n'}
+      Tour Mirabeau{'\n'}
+      39-43 quai André-Citroën{'\n'}
+      75739 Paris Cedex 15{'\n'}
+      France{'\n'}
+      {'\n'}
+      Téléphone : 01 44 38 36 02
     </P>
     <Spacer size={30} />
-    <H2>Directrice de l’application</H2>
+    <H2>Directrice de la publication</H2>
     <Spacer size={30} />
-    <P>Madame Amélie Verdier, Directrice générale de l’Agence régionale de santé Île-de-France.</P>
+    <P>Madame Anne JEANJEAN, Directrice du numérique des ministères sociaux.</P>
     <Spacer size={30} />
-    <H2>Hébergement de l’application</H2>
+    <H2>Hébergement de la Plateforme</H2>
     <Spacer size={30} />
-    <P>L'pplication est hébergée par{'\u00A0'}:</P>
+    <P>Cette plateforme est hébergée par{'\u00A0'}:</P>
     <P>
       OVH SAS{'\n'}2 rue Kellermann{'\n'}
       59100 Roubaix{'\n'}